Create Rich network
I am about to create a new script which will be based on using a graph that contains all the necessary information as part of the meta data. This will require all the functions to be re-written but will probably simplify a lot of operations. However it could break everything so I am going to push and merge this branch then start a new branch