Function Changes
I changed the functions so that the process doesn't crash when the grid collapses. 
I also tried to reduce computation time by only performing the power flow operations on the coponents that needed it. however this didn't really improve things as it takes so much time to calculate which need changeing