Delete Carefully Bug
After updating Delete Carefully so that it can remove edges, I induced a bug as I have made the node part for edges as well. This made deletecarefully not work for nodes. This was a small change and fixed the problem. 
A couple of tidy up or explanations have been added to some related functions.