Extract stats changes
The grid loading metric wasn't as useful as I had realised. I have now added in multiple loading/alpha metrics. This is effectively a breaking change so I have taken advantage of the moment to change all the variable names as the function won't work in old code now anyway